refactor(evm): erc20 precompile config + module config
