--- modules/evm/src/precompiles/erc20.config.json
@@ -1,20 +1,0 @@
-{
-    "abi": [
-        "function transfer(address to, uint256 amount) external",
-        "function mint(address to, uint256 amount) external",
-        "function approve(address spender, uint256 amount) external returns (bool)",
-        "function balanceOf(address account) external view returns (uint256)",
-        "function increaseAllowance(address spender, uint256 addedValue) external returns (bool)",
-        "function allowance(address owner, address spender) external view returns (uint256)",
-        "function burn(uint256 amount) external",
-        "function burn(address account, uint256 amount) external",
-        "function burnFrom(address account, uint256 amount) external",
-        "function transferOwnership(address newOwner) external",
-        "function owner() external view returns (address)",
-        "event Transfer(address indexed from, address indexed to, uint256 value)",
-        "event Approval(address indexed owner, address indexed spender, uint256 value)"
-    ],
-    "contractAddress": "0xD4949664cD82660AaE99bEdc034a0deA8A0bd517",
-    "amount": "1",
-    "feeFund": "10000000"
-}
