+++ modules/evm/src/precompiles/erc20/config/config.ts
@@ -1,0 +1,8 @@
+import { PrecompileConfig } from "../../common/types/precompile.types";
+import { Interface } from "ethers";
+
+export type ERC20PrecompileConfig = PrecompileConfig<{
+    interface: Interface;
+    amount: string;
+    feeFund: string;
+}>;
