+++ modules/evm/test/precompiles/erc20/utils/helpers.ts
@@ -1,5 +1,5 @@
 import { ethers } from "hardhat";
-import { TransactionReceipt, Log, Interface, LogDescription, TransactionResponse, EventLog, Contract, ContractTransaction } from "ethers";
+import { TransactionReceipt, Log, Interface, Contract } from "ethers";
 import { expect } from "chai";
 import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers";
 import { executeTx } from "@testing/hardhat/utils";
